Add stream, update testing.
Add stream function using the direct-to-S3 upload, which is the more
generally usable API endpoint as it can handle large data. Switch all
other file upload functions to act as a wrapper to this. (This change
changes upload_file to be synonymous to upload_aws.)

Switch file upload testing to use example files included in the repo,
and add test for upload_stream.